+++ CHANGELOG.md
@@ -32,3 +32,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
  - Allows for deleted block settings
+
+## [v0.1.7] 2022-07-28
+
+### Added
+ - Email field support
+
+### Fixed
+ - Wrapped url field in link